Refactor JsonSchemaData out of JsonSchemaConverter
